Upgrade to SIS 1.5. This include an upgrade from EPSG 9.9.1 to EPSG 12.026.
Known issue:
The datum element of the WGS84 CRS disapears, because it is now a datum ensemble.
For updating the POJO, we need to either depends on a snapshot version of GeoAPI
or depends directly on Apache SIS 1.5, until a new GeoAPI version is released.
A temporary workaround will be adressed in a separated branch.